refactor: permit edc-build and autodoc to work separately
