+++ .github/workflows/release-edc-build-plugin.yml
@@ -6,7 +6,6 @@ on:
       version:
         required: true
         description: version to be released
-  push: # todo: to be removed
 
 jobs:
   release:
@@ -18,13 +17,26 @@ jobs:
         with:
           gpg-private-key: ${{ secrets.ORG_GPG_PRIVATE_KEY }}
       - run: | 
-          ./gradlew -Pversion=0.0.1 -Psigning.gnupg.executable=gpg -Psigning.gnupg.passphrase="${{ secrets.ORG_GPG_PASSPHRASE }}" \
+          ./gradlew -Pversion=${{ inputs.version }} \
+                    -Psigning.gnupg.executable=gpg \
+                    -Psigning.gnupg.passphrase="${{ secrets.ORG_GPG_PASSPHRASE }}" \
                     :plugins:edc-build:publishPlugins
         env:
           GRADLE_PUBLISH_KEY: ${{ secrets.GRADLE_PUBLISH_KEY }}
           GRADLE_PUBLISH_SECRET: ${{ secrets.GRADLE_PUBLISH_SECRET }}
 
-# todo: create tag, create release
+      - name: Create GitHub Release
+        uses: ncipollo/release-action@v1
+        with:
+          generateReleaseNotes: true
+          tag: edc-build-v${{ inputs.version }}
+          token: ${{ secrets.GITHUB_TOKEN }}
+          removeArtifacts: true
+          makeLatest: true
 
-
-
+      - uses: sarisia/actions-status-discord@v1
+        if: always()
+        with:
+          webhook: ${{ secrets.DISCORD_GITHUB_CI_WEBHOOK }}
+          title: "Release edc-build-v${{ inputs.version }} plugin"
+          username: GitHub Actions
