+++ plugins/edc-build/src/main/java/org/eclipse/edc/plugins/edcbuild/plugins/CheckModuleNamesAction.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.plugins.modulenames;
+package org.eclipse.edc.plugins.edcbuild.plugins;
 
 import org.gradle.api.Action;
 import org.gradle.api.GradleException;
