+++ plugins/edc-build/src/test/java/org/eclipse/edc/plugins/edcbuild/plugins/ModuleNamesPluginTest.java
@@ -12,7 +12,8 @@
  *
  */
 
-package org.eclipse.edc.plugins.modulenames;
+package org.eclipse.edc.plugins.edcbuild.plugins;
+
 
 import org.gradle.api.GradleException;
 import org.gradle.api.Project;
