--- plugins/module-names/README.md
@@ -1,2 +1,0 @@
-This module contains a gradle plugin that verifies that there are no two modules with the same name in a multi-module
-gradle project as this may cause problems during the build.
