--- plugins/openapi-merger/README.md
@@ -1,1 +1,0 @@
-This module contains a gradle plugin that allows to merge partial OpenAPI spec files into one file.
