--- plugins/test-summary/README.md
@@ -1,1 +1,0 @@
-This module contains a gradle plugin that prints a summary about a JUnit test run to the console.
