+++ settings.gradle.kts
@@ -4,7 +4,4 @@ include(":plugins:autodoc:autodoc-plugin")
 include(":plugins:autodoc:autodoc-processor")
 include(":plugins:autodoc:autodoc-converters")
 include(":plugins:edc-build")
-include(":plugins:module-names")
-include(":plugins:openapi-merger")
-include(":plugins:test-summary")
 include(":version-catalog")
